groq_prompt_reverse/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
groq>=0.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
torch>=2.0.0 